+++ op/console/src/main/java/org/jboss/hal/op/endpoint/EndpointTable.java
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.jboss.hal.op.bootstrap;
+package org.jboss.hal.op.endpoint;
 
 import java.util.List;
 import java.util.function.Consumer;